refactor: extract BaseDialog severity header
Part of #103
